Diet-Recommendation-System-main/Streamlit_Frontend/requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.1
numpy==1.24.1
# streamlit==1.16.0
streamlit==1.24.1
streamlit-echarts==0.3.0
requests==2.28.1
altair==4.0
python-dotenv==1.0.0
pyjwt==2.6.0
extra-streamlit-components==0.1.55
bcrypt==4.0.1
streamlit-extras
simplejson==3.17.6
# sqlite3 is built-in with Python, no need to specify
